refactor(commands): extract /doctor command into doctor.rs
